🔥 delete duplicated files that were moved in main.
