+++ airflow/ui/src/pages/DagsList.tsx
@@ -16,92 +16,56 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-
-import { ColumnDef } from "@tanstack/react-table";
-import { useSearchParams } from "react-router-dom";
 import {
   Badge,
-  Button,
-  ButtonProps,
   Checkbox,
   Heading,
   HStack,
-  Input,
-  InputGroup,
-  InputGroupProps,
-  InputLeftElement,
-  InputProps,
-  InputRightElement,
   Select,
   Spinner,
   VStack,
 } from "@chakra-ui/react";
+import type { ColumnDef } from "@tanstack/react-table";
 import { Select as ReactSelect } from "chakra-react-select";
-import { FiSearch } from "react-icons/fi";
+import { type ChangeEventHandler, useCallback } from "react";
+import { useSearchParams } from "react-router-dom";
 
-import { DAG } from "openapi/requests/types.gen";
 import { useDagServiceGetDags } from "openapi/queries";
+import type { DAG } from "openapi/requests/types.gen";
+
 import { DataTable } from "../components/DataTable";
-import { pluralize } from "../utils/pluralize";
 import { useTableURLState } from "../components/DataTable/useTableUrlState";
+import { QuickFilterButton } from "../components/QuickFilterButton";
+import { SearchBar } from "../components/SearchBar";
+import { pluralize } from "../utils/pluralize";
 
-const SearchBar = ({
-  groupProps,
-  inputProps,
-  buttonProps,
-}: {
-  groupProps?: InputGroupProps;
-  inputProps?: InputProps;
-  buttonProps?: ButtonProps;
-}) => (
-  <InputGroup {...groupProps}>
-    <InputLeftElement pointerEvents="none">
-      <FiSearch />
-    </InputLeftElement>
-    <Input placeholder="Search DAGs" pr={150} {...inputProps} />
-    <InputRightElement width={150}>
-      <Button
-        variant="ghost"
-        colorScheme="blue"
-        width={140}
-        height="1.75rem"
-        fontWeight="normal"
-        {...buttonProps}
-      >
-        Advanced Search
-      </Button>
-    </InputRightElement>
-  </InputGroup>
-);
-
-const columns: ColumnDef<DAG>[] = [
+const columns: Array<ColumnDef<DAG>> = [
   {
     accessorKey: "dag_id",
-    header: "DAG",
     cell: ({ row }) => row.original.dag_display_name,
+    header: "DAG",
   },
   {
     accessorKey: "is_paused",
-    header: () => "Is Paused",
     enableSorting: false,
+    header: () => "Is Paused",
   },
   {
     accessorKey: "timetable_description",
-    header: () => "Schedule",
     cell: (info) =>
-      info.getValue() !== "Never, external triggers only"
-        ? info.getValue()
-        : undefined,
+      info.getValue() === "Never, external triggers only"
+        ? undefined
+        : info.getValue(),
     enableSorting: false,
+    header: () => "Schedule",
   },
   {
     accessorKey: "next_dagrun",
-    header: "Next DAG Run",
     enableSorting: false,
+    header: "Next DAG Run",
   },
   {
     accessorKey: "tags",
-    header: () => "Tags",
     cell: ({ row }) => (
       <HStack>
         {row.original.tags?.map((tag) => (
@@ -110,53 +74,67 @@ const columns: ColumnDef<DAG>[] = [
       </HStack>
     ),
     enableSorting: false,
+    header: () => "Tags",
   },
 ];
 
-const QuickFilterButton = ({ children, ...rest }: ButtonProps) => (
-  <Button
-    borderRadius={20}
-    fontWeight="normal"
-    colorScheme="blue"
-    variant="outline"
-    {...rest}
-  >
-    {children}
-  </Button>
-);
-
 const PAUSED_PARAM = "paused";
 
-export const DagsList = () => {
-  const cardView = false;
+// eslint-disable-next-line complexity
+export const DagsList = ({ cardView = false }) => {
   const [searchParams, setSearchParams] = useSearchParams();
 
   const showPaused = searchParams.get(PAUSED_PARAM) === "true";
 
-  const { tableURLState, setTableURLState } = useTableURLState();
-  const { sorting, pagination } = tableURLState;
+  const { setTableURLState, tableURLState } = useTableURLState();
+  const { pagination, sorting } = tableURLState;
 
   // TODO: update API to accept multiple orderBy params
-  const sort = sorting[0];
+  const [sort] = sorting;
   const orderBy = sort ? `${sort.desc ? "-" : ""}${sort.id}` : undefined;
 
   const { data, isLoading } = useDagServiceGetDags({
     limit: pagination.pageSize,
     offset: pagination.pageIndex * pagination.pageSize,
     onlyActive: true,
-    paused: showPaused,
     orderBy,
+    paused: showPaused,
   });
+
+  const handlePausedChange = useCallback(() => {
+    searchParams[showPaused ? "delete" : "set"](PAUSED_PARAM, "true");
+    setSearchParams(searchParams);
+    setTableURLState({ pagination: { ...pagination, pageIndex: 0 }, sorting });
+  }, [
+    pagination,
+    searchParams,
+    setSearchParams,
+    setTableURLState,
+    showPaused,
+    sorting,
+  ]);
+
+  const handleSortChange = useCallback<ChangeEventHandler<HTMLSelectElement>>(
+    ({ currentTarget: { value } }) => {
+      setTableURLState({
+        pagination,
+        sorting: value
+          ? [{ desc: value.startsWith("-"), id: value.replace("-", "") }]
+          : [],
+      });
+    },
+    [pagination, setTableURLState],
+  );
 
   return (
     <>
-      {isLoading && <Spinner />}
-      {!isLoading && !!data?.dags && (
+      {isLoading ? <Spinner /> : undefined}
+      {!isLoading && Boolean(data?.dags) && (
         <>
           <VStack alignItems="none">
             <SearchBar
-              inputProps={{ isDisabled: true }}
               buttonProps={{ isDisabled: true }}
+              inputProps={{ isDisabled: true }}
             />
             <HStack justifyContent="space-between">
               <HStack>
@@ -166,59 +144,38 @@ export const DagsList = () => {
                   <QuickFilterButton isDisabled>Running</QuickFilterButton>
                   <QuickFilterButton isDisabled>Successful</QuickFilterButton>
                 </HStack>
-                <Checkbox
-                  isChecked={showPaused}
-                  onChange={() => {
-                    if (showPaused) searchParams.delete(PAUSED_PARAM);
-                    else searchParams.set(PAUSED_PARAM, "true");
-                    setSearchParams(searchParams);
-                    setTableURLState({
-                      sorting,
-                      pagination: { ...pagination, pageIndex: 0 },
-                    });
-                  }}
-                >
+                <Checkbox isChecked={showPaused} onChange={handlePausedChange}>
                   Show Paused DAGs
                 </Checkbox>
               </HStack>
-              <ReactSelect placeholder="Filter by tag" isDisabled />
+              <ReactSelect isDisabled placeholder="Filter by tag" />
             </HStack>
             <HStack justifyContent="space-between">
               <Heading size="md">
-                {pluralize("DAG", data.total_entries)}
+                {pluralize("DAG", data?.total_entries)}
               </Heading>
-              {cardView && (
+              {cardView ? (
                 <Select
+                  onChange={handleSortChange}
                   placeholder="Sort by…"
-                  width="200px"
-                  variant="flushed"
                   value={orderBy}
-                  onChange={(e) => {
-                    setTableURLState({
-                      sorting: e.target.value
-                        ? [
-                            {
-                              id: e.target.value.replace("-", ""),
-                              desc: e.target.value.startsWith("-"),
-                            },
-                          ]
-                        : [],
-                      pagination,
-                    });
-                  }}
+                  variant="flushed"
+                  width="200px"
                 >
                   <option value="dag_id">Sort by DAG ID (A-Z)</option>
                   <option value="-dag_id">Sort by DAG ID (Z-A)</option>
                 </Select>
+              ) : (
+                false
               )}
             </HStack>
           </VStack>
           <DataTable
-            data={data.dags}
-            total={data.total_entries}
             columns={columns}
+            data={data?.dags ?? []}
             initialState={tableURLState}
             onStateChange={setTableURLState}
+            total={data?.total_entries}
           />
         </>
       )}
